Create edit_2
I learnt that the little spikes it from me increasing the stroke weight and not increasing the theta += that is in charge of the amount of lines/dots around the shape. I just had to play around with trying to find the sweet spot. however, I could find the answer to why one line, in particular, is lagging.